Moved VolatileTuple to top level and renamed Tuple.
