+++ lib/concurrent/thread_safe/util.rb
@@ -10,12 +10,12 @@ module Concurrent
       MAX_INT         = (2 ** FIXNUM_BIT_SIZE) - 1
       CPU_COUNT       = 16 # is there a way to determine this?
 
+      autoload :Tuple,           'concurrent/tuple'
       autoload :Volatile,        'concurrent/volatile'
       autoload :Adder,           'concurrent/thread_safe/util/adder'
       autoload :CheapLockable,   'concurrent/thread_safe/util/cheap_lockable'
       autoload :PowerOfTwoTuple, 'concurrent/thread_safe/util/power_of_two_tuple'
       autoload :Striped64,       'concurrent/thread_safe/util/striped64'
-      autoload :VolatileTuple,   'concurrent/thread_safe/util/volatile_tuple'
       autoload :XorShiftRandom,  'concurrent/thread_safe/util/xor_shift_random'
     end
   end
