+++ lib/concurrent/thread_safe/util/power_of_two_tuple.rb
@@ -1,3 +1,5 @@
+require 'concurrent/tuple'
+
 module Concurrent
 
   # @!visibility private
@@ -7,7 +9,7 @@ module Concurrent
     module Util
       
       # @!visibility private
-      class PowerOfTwoTuple < VolatileTuple
+      class PowerOfTwoTuple < Concurrent::Tuple
 
         def initialize(size)
           raise ArgumentError, "size must be a power of 2 (#{size.inspect} provided)" unless size > 0 && size & (size - 1) == 0
