--- lib/concurrent/volatile_tuple.rb
@@ -1,55 +1,0 @@
-module Concurrent
-
-  # @!visibility private
-  module ThreadSafe
-
-    # @!visibility private
-    module Util
-
-      # A fixed size array with volatile volatile getters/setters.
-      # Usage:
-      #   arr = VolatileTuple.new(16)
-      #   arr.volatile_set(0, :foo)
-      #   arr.volatile_get(0)    # => :foo
-      #   arr.cas(0, :foo, :bar) # => true
-      #   arr.volatile_get(0)    # => :bar
-      #
-      # @!visibility private
-      class VolatileTuple
-        include Enumerable
-
-        Tuple = defined?(Rubinius::Tuple) ? Rubinius::Tuple : Array
-
-        def initialize(size)
-          @tuple = tuple = Tuple.new(size)
-          i = 0
-          while i < size
-            tuple[i] = Concurrent::AtomicReference.new
-            i += 1
-          end
-        end
-
-        def volatile_get(i)
-          @tuple[i].get
-        end
-
-        def volatile_set(i, value)
-          @tuple[i].set(value)
-        end
-
-        def compare_and_set(i, old_value, new_value)
-          @tuple[i].compare_and_set(old_value, new_value)
-        end
-        alias_method :cas, :compare_and_set
-
-        def size
-          @tuple.size
-        end
-
-        def each
-          @tuple.each {|ref| yield ref.get}
-        end
-      end
-    end
-  end
-end
